+++ src/game/tests.rs
@@ -524,6 +524,7 @@ mod banding_assignment;
 mod banding_blocked;
 mod banding_formation;
 mod banding_with_other;
+mod blaze_of_glory;
 mod blocking_after_death;
 mod blocking_prevention;
 mod blocking_relation;